+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Make `AcceptedContext::input` and `AcceptedContext::output` available in `const` context.
+- Add `Combinator::log` to debug combinators.
 
 ## [0.6.0] - 2025-01-31